% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docGetClusteringLabelsForEachSection.R,
%   R/getClusteringLabelsForEachSection.R
\docType{package}
\name{getClusteringLabelsForEachSection}
\alias{getClusteringLabelsForEachSection}
\title{Scanpath Aggregation Package}
\usage{
getClusteringLabelsForEachSection(
  convex_hull,
  OR_p,
  compareFirstAndRest = FALSE
)
}
\arguments{
\item{convex_hull}{output of the function getConcaveHulls}

\item{OR_p}{numeric between 0 and 1, threshold of the similarity of two polygons}

\item{compareFirstAndRest}{logic with default FALSE, whether to compare the the polygon on
first section with polygons on the rest sections within a short time AOI}
}
\value{
A vector of integers indicating the clustering number of each section
}
\description{
clutering the cross sections and get clustering labels for each cross section
}
\examples{
data(scanpath_MIT1003_i1182314083)
concave_hull = getConcaveHulls(scanpath_MIT1003_i1182314083, 3, 0.04)
clusterings = getClusteringLabelsForEachSection(concave_hull, 0.2)
}
\author{
He Huang
}
