% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docGetConcaveHulls.R, R/getConcaveHulls.R
\docType{package}
\name{getConcaveHulls}
\alias{getConcaveHulls}
\title{Scanpath Aggregation Package}
\usage{
getConcaveHulls(Fixation_singleStimulus, ExperimentDuration, et, eps_ini = 400)
}
\arguments{
\item{Fixation_singleStimulus}{data.frame containing the following 6 column, including
(1) Participant, which is the name or id for the participants
(2) Fixation_Position_X_px
(3) Fixation_Position_Y_px
(4) Event_Start_Trial_Time_ms, which is the start time of fixation in seconds
(5) Event_Ende_Trial_Time_ms, which is end time of the fixation in seconds
(6) Duration, duration of the fixation in seconds.
Each row of the data.frame is a fixation.}

\item{ExperimentDuration}{numeric, total time of viewing in seconds}

\item{et}{numeric,distance of the cross sections in seconds.}

\item{eps_ini}{numeric, the initial epsilon of DBSCAN, which should be half of stimulus width in pixel}
}
\value{
A list of Lists with one element, wchich ist a numeric matrix of two column.
The output of this function can be uses direct in the function getClusteringLabelsForEachSection()
}
\description{
Calculate concave hulls on each cross section
}
\examples{
data(scanpath_MIT1003_i1182314083)
getConcaveHulls(scanpath_MIT1003_i1182314083, 3, 0.04)
}
\author{
He Huang
}
