% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docGetRepresentScanpath.R,
%   R/getRepresentScanpath.R
\docType{package}
\name{getRepresentScanpath}
\alias{getRepresentScanpath}
\title{Scanpath Aggregation Package}
\usage{
getRepresentScanpath(
  concave_hull,
  clusterings,
  et,
  Fixation_singleStimulus = NULL,
  unionAOI = TRUE,
  durationrec = FALSE
)
}
\arguments{
\item{concave_hull}{output of the function getConcaveHulls()}

\item{clusterings}{output of the function getClusteringLabelsForEachSection()}

\item{et}{numeric,distance of the cross sections in seconds.}

\item{Fixation_singleStimulus}{data.frame with default NULL containing 6 column, including
(1) Participant, which is the name or id for the participants
(2) Fixation_Position_X_px
(3) Fixation_Position_Y_px
(4) Event_Start_Trial_Time_ms, start time of fixation in seconds
(5) Event_Ende_Trial_Time_ms, end time of the fixation in seconds
(6) Duration
Only needed by durationrec = TRUE, otherwise use the default value NULL.}

\item{unionAOI}{logical with default TRUE, using the Union (TRUE) of intersection(FALSE) of the polygons as the shape of the short time AOIs}

\item{durationrec}{logical, if to recalculate the duration of the fixation in representative scanpath using the
average duration of the fixations in the short time AOI}
}
\value{
A list of two elements:
 -representScanpath: a data.frame for the representScanpath
 -AOIpolygons: List of elements of class 'SpatialPolygons' from R package "sp"
}
\description{
Function to find a representative scanpath by giving concave polygons (AOIs) on each cross section and the
clustering number of each cross section
}
\examples{
data(scanpath_MIT1003_i1182314083)
concave_hull = getConcaveHulls(scanpath_MIT1003_i1182314083, 3, 0.04)
clusterings = getClusteringLabelsForEachSection(concave_hull, 0.2)
RepresentScanpath = getRepresentScanpath(concave_hull, clusterings, 0.04,
                                            scanpath_MIT1003_i1182314083)
}
\author{
He Huang
}
