% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docExampleDataMIT1003.R
\docType{data}
\name{scanpath_MIT1003_i1182314083}
\alias{scanpath_MIT1003_i1182314083}
\title{Data to show the examples}
\format{
A data frame with 128 rows and 7 variables
 \describe{
     \item{Participant}{paticipant id}
     \item{Stimulus}{Stimulus id}
     \item{Fixation_Position_X_px}{Fixation_Position_X_px}
     \item{Fixation_Position_Y_px}{Fixation_Position_Y_px}
     \item{Event_Start_Trial_Time_ms}{start time of fixation in seconds}
     \item{Event_Ende_Trial_Time_ms}{end time of fixation in seconds}
     \item{Duration}{duration in seconds}
 }

 @source {data set MIT1003 from Judd et al. (2009)  <Judd, T., Ehinger, K., Durand, F., & Torralba, A. (2009).
  Learning to predict where humans look. 2106–2113.> }
 @examples
 data(scanpath_MIT1003_i1182314083)
}
\usage{
scanpath_MIT1003_i1182314083
}
\description{
A group of scanpaths for one of the stimuli in MIT1003 data  from Judd et al. (2009)
}
\keyword{datasets}
