% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docExampleDataOSIE.R
\docType{data}
\name{scanpath_OSIE_1001}
\alias{scanpath_OSIE_1001}
\title{Data to show the examples contain no start and end time of fixation}
\format{
A data frame with 141 rows and 4 variables
 \describe{
     \item{Participant}{paticipant id}
     \item{Fixation_Position_X_px}{Fixation_Position_X_px}
     \item{Fixation_Position_Y_px}{Fixation_Position_Y_px}
     \item{Duration}{duration in seconds}
 }

 @source {Xu, Jiang, Wang, Kankanhalli, & Zhao (2014)  <Xu, J., Jiang, M., Wang, S., Kankanhalli, M. S., & Zhao, Q. (2014).
 Predicting human gaze beyond pixels. Journal of Vision, 14(1), 28–28.>}
 @examples
 data(scanpath_OSIE_1001)
}
\usage{
scanpath_OSIE_1001
}
\description{
A group of scanpaths for one of the stimululi in OSIE data from Xu, Jiang, Wang, Kankanhalli, & Zhao (2014).
}
\keyword{datasets}
