% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docScanpathAggregation.R,
%   R/scanpathAggregation.R
\docType{package}
\name{scanpathAggr}
\alias{scanpathAggr}
\title{Scanpath Aggregation Package}
\usage{
scanpathAggr(
  scanpaths,
  ExperimentDuration,
  et = 0.04,
  theta = 0.2,
  simplify = FALSE,
  alpha = 50
)
}
\arguments{
\item{scanpaths}{data.frame containing the following 4 or 6 column, including
(1) Participant, which is the name or id for the participants
(2) Fixation_Position_X_px
(3) Fixation_Position_Y_px
(4) Event_Start_Trial_Time_ms, start time of fixation in seconds
(5) Event_Ende_Trial_Time_ms, end time of the fixation in seconds
(6) Duration
If there are no (4) and (5), this function will create fake start and end time for the fixations.
Each row represents a fixation, and fixations are  in temporal order.}

\item{ExperimentDuration}{numeric, total time of viewing in seconds}

\item{et}{numeric with default 0.04,time distance of the cross sections in seconds.}

\item{theta}{numeric between 0 and 1, threshold of the similarity of two polygons}

\item{simplify}{logical with default FALSE, if TRUE, the found representative scanpath
will be simplified by giving threshold alpha in pixel}

\item{alpha}{numeric with defaulf 50 (pixel), threshold of the simplification in pixel,
fixations in representative scanpath closer than 50 pixel will be merged through their AOIs}
}
\value{
A list of two elements:
 -representScanpath: a data.frame for the representScanpath
 -AOIpolygons: List of elements of class 'SpatialPolygons' from R package "sp"
}
\description{
Aggregate a group of scanpaths by finding a representative scanpath based on short time AOIs.
}
\examples{
data(scanpath_MIT1003_i1182314083)
scanpathAggr(scanpath_MIT1003_i1182314083, 3)
}
\author{
He Huang
}
