% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docScanpathVisualisation.R,
%   R/scanpathVisualisation.R
\docType{package}
\name{scanpathVisualisation}
\alias{scanpathVisualisation}
\title{Scanpath Aggregation Package}
\usage{
scanpathVisualisation(
  scanPath = NULL,
  representScanpath = NULL,
  stimulusImgPath,
  drawAOIs = TRUE
)
}
\arguments{
\item{scanPath}{data.frame with default NULL containing 6 column, including
(1) Participant, which is the name or id for the participants
(2) Fixation_Position_X_px
(3) Fixation_Position_Y_px
(4) Event_Start_Trial_Time_ms, start time of fixation in seconds
(5) Event_Ende_Trial_Time_ms, end time of the fixation in seconds
(6) Duration
These are the original scanpaths, form which the representative scanpath is computed.
By NULL draw this function only the representative scanpath}

\item{representScanpath}{The obtained representative scanpath, output of the function getReprentativeScanpath() or scanpathAggr().
The default ist NULL, under which show draw this function only the original scanpaths.}

\item{stimulusImgPath}{file path of the stimulus}

\item{drawAOIs}{logical (default: TRUE), whether to draw the short time AOIs contained in the representScanpath}
}
\value{
A object of class ggplot from the package ggplot2
}
\description{
Scanpath Visualisation
}
\examples{
data(scanpath_MIT1003_i1182314083)
reprensentScanpath = scanpathAggr(scanpath_MIT1003_i1182314083, 3)
stimuluspath = system.file('stimulus/i1182314083.jpeg', package = "scanpathAggr")
scanpathVisualisation(scanpath_MIT1003_i1182314083, reprensentScanpath, stimuluspath)
}
\author{
He Huang
}
